+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 3.15.1 -- 2022-11-14
+
+### Added
+
+* `extendedAdaClass` and `pextendedAdaClass` constants for convenience.
+* `isExtendedAdaClass` to quickly determine whether we have the ADA extended
+  asset class, along with its Plutarch equivalent `pisExtendedAdaClass`.
+
 ## 3.15.0 -- 2022-11-11
 
 ### Added